+++ seneca.js
@@ -250,6 +250,8 @@ const option_defaults = {
         rules: false,
         // If false, use Gubu for option validation (including plugin defaults)
         options: true,
+        // If true, look for plugin options by name at the top level of options
+        top_plugins: false,
     }),
     // Processing task ordering.
     order: {